requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
openai==1.3.8
python-dotenv==1.0.0
rich==13.7.0
typer==0.9.0
httpx==0.25.2
asyncio-mqtt==0.16.1
